itest: add new file `lnd_bump_fee.go`
Move bumpfee RPC related tests into one file - these tests focus on
testing the behaivor of the RPC only without any force close context.
